requirements.txt
python_chess == 0.30.1
